+++ thprac/src/thprac/thprac_gui_input.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 namespace THPrac
 {
@@ -12,7 +13,7 @@ namespace THPrac
 				INGAGME_INPUT_GEN1,
 				INGAGME_INPUT_GEN2,
 		};
-		bool InGameInputInit(ingame_input_gen_t gen, int reg1, int reg2, int reg3 = 0);
+		bool InGameInputInit(ingame_input_gen_t gen, uintptr_t reg1, uintptr_t reg2, uintptr_t reg3 = 0);
 		bool InGameInputGet(int key);
 
 		// Keyboard Input
